Complete Decimal migration for scenarios, coach, forecasting, and history domains
Converts remaining money fields from float to Decimal across the scenario
planning, coaching, forecasting, and history subsystems, matching the
earlier core-domain migration. Percentages, scores, and ratios (savings
rate, debt-to-income, risk/sustainability scores) intentionally remain
float, converted from Decimal only at the display boundary.

Also fixes two latent bugs surfaced by the migration: to_money() raised
uncatchable decimal.InvalidOperation instead of ValueError on invalid
strings, and ScenarioImpact/ScenarioComparison money-JSON round-tripping
was missing.

+++ src/financial/coach/explanation.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from src.financial.coach.models import (
     AdviceExplanation,
     CoachingAdvice,
@@ -12,7 +14,7 @@ from src.financial.scenarios.report import (
 
 def _format_change(
     metric: str,
-    change: float,
+    change: Decimal,
 ) -> str:
     """Format one projected financial change."""
     if metric == "Health Score":
